Separate transcript message component
